healthcare-mcp-public/requirements.txt
fastapi==0.115.12
mcp==1.6.0
pydantic==2.11.3
pytest==8.3.5
python-dotenv==1.1.0
Requests==2.32.3
slowapi==0.1.9
structlog==25.2.0
uvicorn==0.34.2
pytest-asyncio
